simplify example spec file also remove the mention of sky and bskyb (helper should be project agnostic)
