+++ component-structure/test/specs/main.spec.js
@@ -1,14 +1,8 @@
-/* Add Assets: HTML + CSS to setup page for functional testing */
-    document.body.innerHTML = __html__['_site/index.html'];
-    function appendCSS(fileObj){
-        var  link = document.createElement('link'); link.rel = 'stylesheet'; link.href='base/' + fileObj.path;  document.body.appendChild(link)
-    }
-    appendCSS({path: '_site/styles/demo.css'});
-    appendCSS({path: '_site/styles/{{ component }}.css'});
-/* End adding assets */
+/* Add HTML + CSS to setup page for functional testing */
+require('../helper').loadAssets();
 
 /* Require file to test */
-var {{ component }} = require('../../src/scripts/{{ component }}');
+var {{ component }} = require('src/scripts/{{ component }}');
 
 /* Start Test */
 describe('{{ component }} module can ', function () {
